worker/requirements.txt
openai-whisper==20240930

pytest==8.3.3
pytest-cov==6.0.0
pika==1.3.2
click==8.1.7
ffmpeg-python==0.2.0
pydantic==2.10.2
requests==2.32.3
pydantic-settings==2.7.1
